move zoom_level out of scope
